refactor(challenge): 미션타입 폴더를 my-challenge 경계 안으로 (도메인-내-도메인 정리, A안)
bonus/ot/daily-mission은 모두 my-challenge에서만 쓰이는 하위기능인데 최상위에
나와 있어 경계가 샜음. my-challenge 안으로 이동해 "삭제 가능 단위"를 명확히:
- challenge/{bonus-mission,ot-mission,daily-mission}/ → my-challenge/ 아래로
- daily-mission 통합: 흩어져 있던 top-level daily-mission + section/daily-mission을
  my-challenge/daily-mission/ 하나로 병합
- 깨진 상대경로는 @/ 절대경로로 갱신(향후 이동에도 안전)

결과: 최상위 challenge가 api/ui/hooks/dashboard/inquiry/my-challenge/old로 축소.
tsc 0 / lint 통과. 동작 변경 없음.
