+++ apps/web/src/domain/challenge/my-challenge/daily-mission/DailyMissionInfoSection.tsx
@@ -1,8 +1,8 @@
 import HybridLink from '@/common/HybridLink';
 import { UserChallengeMissionDetail } from '@/schema';
 import { BONUS_MISSION_TH } from '@/utils/constants';
-import OtVideo from '../../../ot-mission/OtVideo';
-import ContentsDropdown from '../../dropdown/ContentsDropdown';
+import OtVideo from '@/domain/challenge/my-challenge/ot-mission/OtVideo';
+import ContentsDropdown from '@/domain/challenge/my-challenge/dropdown/ContentsDropdown';
 
 interface Props {
   missionDetail: UserChallengeMissionDetail;
