+++ apps/web/src/domain/challenge/my-challenge/daily-mission/DailyMissionSection.tsx
