+++ apps/web/src/domain/challenge/my-challenge/mission/AbsentMissionDetailMenu.tsx
@@ -1,8 +1,8 @@
 import { Schedule, UserChallengeMissionDetail } from '@/schema';
 import { BONUS_MISSION_TH } from '@/utils/constants';
-import BonusMissionSubmitMenu from '../../bonus-mission/BonusMissionSubmitMenu';
-import OtMissionSubmitMenu from '../../ot-mission/OtMissionSubmitMenu';
-import OtVideo from '../../ot-mission/OtVideo';
+import BonusMissionSubmitMenu from '@/domain/challenge/my-challenge/bonus-mission/BonusMissionSubmitMenu';
+import OtMissionSubmitMenu from '@/domain/challenge/my-challenge/ot-mission/OtMissionSubmitMenu';
+import OtVideo from '@/domain/challenge/my-challenge/ot-mission/OtVideo';
 import MenuContentsDropdown from '../dropdown/MenuContentsDropdown';
 import AbsentContentsInfoMenu from '../menu/AbsentContentsInfoMenu';
 import AbsentMissionSubmitMenu from '../menu/AbsentMissionSubmitMenu';
