+++ apps/web/src/domain/challenge/my-challenge/mission/DoneMissionDetailMenu.tsx
@@ -7,7 +7,7 @@ import {
 } from '@/schema';
 import { BONUS_MISSION_TH } from '@/utils/constants';
 import { useParams } from 'next/navigation';
-import OtVideo from '../../ot-mission/OtVideo';
+import OtVideo from '@/domain/challenge/my-challenge/ot-mission/OtVideo';
 import MenuContentsDropdown from '../dropdown/MenuContentsDropdown';
 import ParsedCommentBox from '../ParsedCommentBox';
 interface Props {
