+++ apps/web/src/domain/challenge/my-challenge/ot-mission/OtMissionInputSection.tsx
